LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.7_Punctuators/S7.7_A2_T5-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S7.7_A2_T5

PASS Expected parsing failure

TEST COMPLETE
